testing/requirements.txt
llama-index>=0.10.0
openai>=1.0.0
python-dotenv>=1.0.0
llama-index-llms-openrouter>=0.3.2 